requirements.txt
selenium
beautifulsoup4
webdriver-manager
python-dotenv